+++ docs/reference/permissions.md
@@ -1,0 +1,3 @@
+# tanjun.permissions
+
+::: tanjun.permissions
