+++ docs/reference/utilities.md
@@ -1,3 +1,3 @@
 # tanjun.utilities
 
-::: tanjun.utilities
+`tanjun.utilities` has been deprecated in favour of [tanjun.permissions][].
