+++ tests/commands/test_menu.py
@@ -42,6 +42,7 @@ import hikari
 import pytest
 
 import tanjun
+from tanjun import _internal
 
 
 def test_as_message_menu():
@@ -402,7 +403,7 @@ class TestMenuCommand:
             .add_check(mock_other_callback)
         )
 
-        with mock.patch.object(tanjun.utilities, "gather_checks", new=mock.AsyncMock()) as gather_checks:
+        with mock.patch.object(_internal, "gather_checks", new=mock.AsyncMock()) as gather_checks:
             result = await command.check_context(mock_context)
 
             gather_checks.assert_awaited_once_with(mock_context, [mock_callback, mock_other_callback])
