+++ tests/commands/test_message.py
@@ -45,6 +45,7 @@ import hikari
 import pytest
 
 import tanjun
+from tanjun import _internal
 from tanjun.commands import base as base_command
 
 _T = typing.TypeVar("_T")
@@ -380,7 +381,7 @@ class TestMessageCommand:
             .add_check(mock_other_callback)
         )
 
-        with mock.patch.object(tanjun.utilities, "gather_checks", new=mock.AsyncMock()) as gather_checks:
+        with mock.patch.object(_internal, "gather_checks", new=mock.AsyncMock()) as gather_checks:
             result = await command.check_context(mock_context)
 
             gather_checks.assert_awaited_once_with(mock_context, [mock_callback, mock_other_callback])
