+++ tests/test_utilities.py
@@ -33,244 +33,21 @@
 # pyright: reportPrivateUsage=none
 # This leads to too many false-positives around mocks.
 
-import typing
-from unittest import mock
-
-import pytest
-
 import tanjun
-from tanjun import utilities
 
-_T = typing.TypeVar("_T")
 
-
-@pytest.mark.asyncio()
-async def test_gather_checks_handles_no_checks():
-    mock_ctx = mock.AsyncMock()
-    assert await utilities.gather_checks(mock_ctx, ()) is True
-
-    mock_ctx.call_with_async_di.assert_not_called()
-
-
-@pytest.mark.asyncio()
-async def test_gather_checks_handles_failed_check():
-    mock_ctx = mock.Mock()
-    mock_ctx.call_with_async_di = mock.AsyncMock(side_effect=[True, False, True])
-    check_1 = mock.Mock()
-    check_2 = mock.Mock()
-    check_3 = mock.Mock()
-
-    assert await utilities.gather_checks(mock_ctx, (check_1, check_2, check_3)) is False
-
-    mock_ctx.call_with_async_di.assert_has_awaits(
-        [mock.call(check_1, mock_ctx), mock.call(check_2, mock_ctx), mock.call(check_3, mock_ctx)]
-    )
-
-
-@pytest.mark.asyncio()
-async def test_gather_checks_handles_check_failed_by_raise():
-    mock_ctx = mock.Mock()
-    mock_ctx.call_with_async_di = mock.AsyncMock(side_effect=[True, tanjun.FailedCheck, True])
-    check_1 = mock.Mock()
-    check_2 = mock.Mock()
-    check_3 = mock.Mock()
-
-    assert await utilities.gather_checks(mock_ctx, (check_1, check_2, check_3)) is False
-
-    mock_ctx.call_with_async_di.assert_has_awaits(
-        [mock.call(check_1, mock_ctx), mock.call(check_2, mock_ctx), mock.call(check_3, mock_ctx)]
-    )
-
-
-@pytest.mark.asyncio()
-async def test_gather_checks():
-    mock_ctx = mock.Mock()
-    mock_ctx.call_with_async_di = mock.AsyncMock(side_effect=[True, True, True])
-    check_1 = mock.Mock()
-    check_2 = mock.Mock()
-    check_3 = mock.Mock()
-
-    assert await utilities.gather_checks(mock_ctx, (check_1, check_2, check_3)) is True
-
-    mock_ctx.call_with_async_di.assert_has_awaits(
-        [mock.call(check_1, mock_ctx), mock.call(check_2, mock_ctx), mock.call(check_3, mock_ctx)]
-    )
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_resource():
-    ...
-
-
-@pytest.mark.parametrize(
-    ("content", "prefix", "expected_result"),
-    [
-        ("no go sir", ("no", "home", "blow"), "no"),
-        ("hime", ("hi", "hime", "boomer"), "hime"),
-        ("boomer", ("boo", "boomer", "no u"), "boomer"),
-        ("ok boomer", ("no", "nani"), None),
-        ("", ("nannnnni",), None),
-        ("ok ok ok", (), None),
-    ],
-)
-def test_match_prefix_names(content: str, prefix: str, expected_result: typing.Optional[str]):
-    assert utilities.match_prefix_names(content, prefix) == expected_result
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_permissions():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_permissions_when_guild_owner():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_permissions_when_admin_role():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_permissions_when_no_channel():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_when_guild_owner():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_when_admin_role():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_when_no_channel():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_when_channel_object_provided():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_for_uncached_entities():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_permissions_for_no_cache():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_everyone_permissions():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_everyone_permissions_admin_role():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-def test_calculate_everyone_permissions_no_channel():
-    ...
-
-
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions_admin_role():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions_for_uncached_entities():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions_for_no_cache():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions_no_channel():
-    ...
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.asyncio()
-async def test_fetch_everyone_permissions_channel_object_provided():
-    ...
-
-
-class TestCastedView:
-    def test___getitem___for_non_existant_entry(self):
-        mock_cast = mock.Mock()
-        view = utilities.CastedView[typing.Any, typing.Any, typing.Any]({}, mock_cast)
-
-        with pytest.raises(KeyError):
-            view["foo"]
-
-        mock_cast.assert_not_called()
-
-    def test___getitem___for_buffered_entry(self):
-        mock_cast = mock.Mock()
-        mock_value = mock.MagicMock()
-        view = utilities.CastedView[typing.Any, str, typing.Any]({"a": "b"}, mock_cast)
-        view._buffer["a"] = mock_value
-
-        result = view["a"]
-
-        assert result is mock_value
-        mock_cast.assert_not_called()
-
-    def test___getitem___for_not_buffered_entry(self):
-        mock_cast = mock.Mock()
-        mock_value = mock.MagicMock()
-        view = utilities.CastedView[typing.Any, typing.Any, typing.Any]({"a": mock_value}, mock_cast)
-
-        result = view["a"]
-
-        assert result is mock_cast.return_value
-        assert view._buffer["a"] is mock_cast.return_value
-        mock_cast.assert_called_once_with(mock_value)
-
-    def test___iter__(self):
-        mock_iter = mock.Mock(return_value=iter((1, 2, 3)))
-        mock_dict = mock.Mock(__iter__=mock_iter)
-        view = utilities.CastedView[int, int, typing.Any](mock_dict, mock.Mock())
-
-        assert iter(view) is mock_iter.return_value
-
-    def test___len___(self):
-        mock_dict = mock.Mock(__len__=mock.Mock(return_value=43123))
-        view = utilities.CastedView[int, int, typing.Any](mock_dict, mock.Mock())
-
-        assert len(view) == 43123
+def test_aliases():
+    assert set(tanjun.utilities.__all__) == {
+        "ALL_PERMISSIONS",
+        "DM_PERMISSIONS",
+        "calculate_everyone_permissions",
+        "calculate_permissions",
+        "fetch_everyone_permissions",
+        "fetch_permissions",
+    }
+    assert tanjun.utilities.ALL_PERMISSIONS is tanjun.permissions.ALL_PERMISSIONS
+    assert tanjun.utilities.DM_PERMISSIONS is tanjun.permissions.DM_PERMISSIONS
+    assert tanjun.utilities.calculate_everyone_permissions is tanjun.permissions.calculate_everyone_permissions
+    assert tanjun.utilities.calculate_permissions is tanjun.permissions.calculate_permissions
+    assert tanjun.utilities.fetch_everyone_permissions is tanjun.permissions.fetch_everyone_permissions
+    assert tanjun.utilities.fetch_permissions is tanjun.permissions.fetch_permissions
